D6 - Changed Quantity Data type in database

+++ src/db/sql/Create OrderLines Table.sql
@@ -5,7 +5,7 @@ CREATE TABLE OrderLines
     Id        INT NOT NULL PRIMARY KEY IDENTITY(1,1),
     OrderId   INT NOT NULL,
     ProductId INT NOT NULL,
-    Quantity  INT NOT NULL,
+    Quantity  DECIMAL(10,2) NOT NULL,
 
     FOREIGN KEY (ProductId) REFERENCES Products (Id),
     FOREIGN KEY (OrderId) REFERENCES Orders (Id)
